common: JobInformation ints to float64

+++ common.go
@@ -37,14 +37,14 @@ type JobInformation struct {
 	// File is the file that is the target of the current print job.
 	File FileInformation `json:"file"`
 	// EstimatedPrintTime is the estimated print time for the file, in seconds.
-	EstimatedPrintTime int `json:"estimatedPrintTime"`
+	EstimatedPrintTime float64 `json:"estimatedPrintTime"`
 	// LastPrintTime is the print time of the last print of the file, in seconds.
-	LastPrintTime int `json:"lastPrintTime"`
+	LastPrintTime float64 `json:"lastPrintTime"`
 	// Filament contains Information regarding the estimated filament
 	// usage of the print job.
 	Filament struct {
 		// Length of filament used, in mm
-		Length int `json:"length"`
+		Length float64 `json:"length"`
 		// Volume of filament used, in cm³
 		Volume float64 `json:"volume"`
 	} `json:"filament"`
@@ -78,9 +78,9 @@ type ProgressInformation struct {
 	// from the beginning.
 	FilePosition int `json:"filepos"`
 	// PrintTime is time already spent printing, in seconds
-	PrintTime int `json:"printTime"`
+	PrintTime float64 `json:"printTime"`
 	// PrintTimeLeft is estimate of time left to print, in seconds
-	PrintTimeLeft int `json:"printTimeLeft"`
+	PrintTimeLeft float64 `json:"printTimeLeft"`
 }
 
 // TemperatureState is the printer’s temperature state data.
